tests/unit/requirements.txt
aiobotocore
aiohttp
aiokafka
asyncio
asyncmock
azure-servicebus
dpath
pytest-asyncio
